moved manual auth utilities to auth folder
